Moved NoteGroups specific functionality of ExampleView to a subclass
