+++ src/palette/view/widgets/noteGroups.cpp
@@ -34,7 +34,6 @@
 #include "translation.h"
 
 using namespace mu::palette;
-using namespace mu::notation;
 using namespace mu::engraving;
 
 Score* NoteGroups::createScore(int n, DurationType t, std::vector<Chord*>* chords)
@@ -73,7 +72,7 @@ Score* NoteGroups::createScore(int n, DurationType t, std::vector<Chord*>* chord
     StaffType* st = c.score()->staff(0)->staffType(Fraction(0, 1));
     st->setLines(1);            // single line only
     st->setGenClef(false);      // no clef
-//      st->setGenTimesig(false); // don't display time sig since ExampleView is unable to reflect custom time sig text/symbols
+//    st->setGenTimesig(false); // don't display time sig since NoteGroupsExampleView is unable to reflect custom time sig text/symbols
 
     return c.score();
 }
@@ -101,13 +100,13 @@ NoteGroups::NoteGroups(QWidget* parent)
 
     connect(resetGroups, &QPushButton::clicked, this, &NoteGroups::resetClicked);
 
-    connect(view8, &ExampleView::noteClicked, this, &NoteGroups::noteClicked);
-    connect(view16, &ExampleView::noteClicked, this, &NoteGroups::noteClicked);
-    connect(view32, &ExampleView::noteClicked, this, &NoteGroups::noteClicked);
+    connect(view8, &NoteGroupsExampleView::noteClicked, this, &NoteGroups::noteClicked);
+    connect(view16, &NoteGroupsExampleView::noteClicked, this, &NoteGroups::noteClicked);
+    connect(view32, &NoteGroupsExampleView::noteClicked, this, &NoteGroups::noteClicked);
 
-    connect(view8, &ExampleView::beamPropertyDropped, this, &NoteGroups::beamPropertyDropped);
-    connect(view16, &ExampleView::beamPropertyDropped, this, &NoteGroups::beamPropertyDropped);
-    connect(view32, &ExampleView::beamPropertyDropped, this, &NoteGroups::beamPropertyDropped);
+    connect(view8, &NoteGroupsExampleView::beamPropertyDropped, this, &NoteGroups::beamPropertyDropped);
+    connect(view16, &NoteGroupsExampleView::beamPropertyDropped, this, &NoteGroups::beamPropertyDropped);
+    connect(view32, &NoteGroupsExampleView::beamPropertyDropped, this, &NoteGroups::beamPropertyDropped);
 }
 
 void NoteGroups::setSig(Fraction sig, const Groups& g, const QString& z, const QString& n)
